requirements.txt
fastapi
uvicorn
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
beautifulsoup4
requests
selenium
openai
pydantic
python-dotenv
psycopg2-binary
sqlalchemy
sendgrid
slack-sdk
pandas
jinja2
aiohttp
newspaper3k
linkedin-api